requirements.txt
# Core routing libraries
semantic-router==0.0.20
routellm==0.0.5

# OpenAI/Azure OpenAI
openai>=1.12.0
tiktoken>=0.5.0

# Vector operations
numpy>=1.24.0
scikit-learn>=1.3.0

# Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# API and async
aiohttp>=3.9.0
requests>=2.31.0
python-dotenv>=1.0.0

# Utilities
colorama>=0.4.6
tabulate>=0.9.0
tqdm>=4.66.0

# Performance monitoring
psutil>=5.9.0

# Optional for advanced features
faiss-cpu>=1.7.4  # For optimized vector search
redis>=5.0.0  # For caching